requirements.txt
asgiref==3.10.0
attrs==25.4.0
black==23.9.1
channels==4.0.0
channels-redis==4.1.0
click==8.3.0
Django==5.2.7
django-cors-headers==4.3.1
django-environ==0.12.0
django-filter==23.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.26.5
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
msgpack==1.1.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.11.1
PyYAML==6.0.3
redis==6.4.0
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
